GLES: Move depth packing code to dedicated file.
